Add error handling for user, product, and comment updates and deletions

+++ backend/src/db/queries.ts
@@ -7,7 +7,7 @@ import {
     type NewUser,
     type NewComment,
     type NewProduct,
-    commentsRelations,
+
 } from "./schema";
 
 //user queries
@@ -21,19 +21,25 @@ export const getUserById = async (id: string) => {
 };
 
 export const updateUser = async (id: string, data: Partial<NewUser>) => {
+    const existingUser = await getUserById(id);
+    if (!existingUser) {
+        throw new Error(`User with ${id} not found`)
+    }
     const [user] = await db
         .update(users)
         .set(data)
         .where(eq(users.id, id))
         .returning();
-    return users;
+    return user;
 };
 //upsert will either create or update
 export const upsertUser = async (data: NewUser) => {
-    const existingUser = await getUserById(data.id);
-    if (existingUser) return updateUser(data.id, data);
-    return createUser(data);
-};
+    const [user] = await db.insert(users).values(data).onConflictDoUpdate({
+        target: users.id,
+        set: data,
+    }).returning();
+    return user;
+}
 
 
 
@@ -73,6 +79,10 @@ export const getProductByUserId = async (userId: string) => {
 };
 
 export const updateProduct = async (id: string, data: Partial<NewProduct>) => {
+    const existingProduct = await getProductById(id);
+    if (!existingProduct) {
+        throw new Error(`Product with ${id} not found`)
+    }
     const [product] = await db
         .update(products)
         .set(data)
@@ -89,6 +99,10 @@ export const createComment = async (data: NewComment) => {
 }
 
 export const deleteComment = async (id: string) => {
+    const existingComment = await getCommentById(id);
+    if (!existingComment) {
+        throw new Error(`Comment with ${id} not found`)
+    }
     const [comment] = await db.delete(comments).where(eq(comments.id, id)).returning();
     return comment;
 }
